Generate and sign the metadata offline
This allows us to fix a few bugs, all at the same time:

 * Ensure *all* files are uploaded to the CDN even if the page load timed out
 * Moving firmwares between different targets is now fast
 * Make the embargo targets look visually different for each vendor

As the different embargo remotes have different names we can use a 'dirty'
indicator to each remote that needs re-generating and signing. It is expected
that a cron job will call generate-metadata.py running as the correct user.

As a side effect of this patch, we also now prevent direct upload to testing
which we also need for the future delayed-signing of firmware.