Sign the firmware files in the cron job
This allows us to move the signing out of the Uploadedfile object, which means
the firmware upload does not block waiting for the plugins to run.

This also adds a new timestamp to the firmware object, which allows us to see
when the firmware was signed. This allows us to veto specific actions, for
example moving it to the public repos.