Only show the legal agreement once rather than on every upload
Also, add the ability to version the agreements and only prompt users of the
LVFS to re-accept the terms when the version number changes.

We need this for the migration to the Linux Foundation, where users will have
to agree that a different legal entity is going to be distributing thier
firmware. The LF legal team also want to polish up the current agreement.