Allow undeleting firmware
Instead of deleting the firmware file we move it to a private location. It is
also no longer included in the metadata or in the search results.

This allows us to preserve the database history for audit reasons, and also to
implement an 'undelete' action.

Fixes https://github.com/hughsie/lvfs-website/issues/89