Allow components to have a priority
This allows us to set the priority of a specific component to be largest, which
allows fwupd to order the results with the 'main' entity listed first.

The other way to do this is to simply add the files into the source .cab file
in a specific order, but I think that's fragile. It's better to be explicit
rather than relying on an assumption that the file order won't be changed when
the archive is repacked.