Only allow editing update descriptions when in private or embargo
This stops vendors uploading a firmware with a good update description, moving
the firmware to stable, and then changing the description be garbage.

The bad update description will have "problems" but as it's already in stable
the LVFS doesn't auto-demote it...

This feels somewhat like cat and mouse at times. Can't all we just care about
what the user reads?!