Make the action more obvious when managing vendor users
On the LVFS we usually put the 'action widget' as a button on the right hand
side of a table. In the per-vendor users view the common action was actually
going to the details, which meant you had to know you could click on the email
address. The uncommon action was a big red scary 'Disable' button that made it
look even more like this was the only way to interact with the table row.

Use a less scary 'Details' button like the rest of the LVFS, and remove the
one-click-disable misfeature.