Perform 'offline' runtime checks on uploaded firmware
Initially, add a simple plugin that checks the UEFI capsule header. Later we
could add further firmware analysis like uefi-firmware-parser or perhaps even
something like CHIPSEC.

Running this as a cron job every 5 minutes means the main uWSGI thread is not
waiting for the result, which means we could even do a malware scan which takes
a minute or two to complete.

Failures in the assay means a problem is added to the firmware, which prevents
it being moved to testing and stable. Some failures are waivable by a QA user
which means a problem is no longer generated.

Fixes https://github.com/hughsie/lvfs-website/issues/227