Add a component property to define the firmware category
Some firmwares only update one part of the system, e.g. the EC or ME firmware.
Other updates include all the updates needed for the whole system, and vendors
have been doing different things with the component name due to this.

Some examples of the brokenness:

* `Logitech Unifying Receiver` -> `Logitech Logitech Unifying Receiver`
* `Thunderbolt Firmware Update for AKiTiO Node Lite` -> `AKiTiO Thunderbolt Firmware Update for AKiTiO Node Lite`
* `Star LabTop Mk III Embedded Controller Firmware` -> `Star Labs Star LabTop Mk III Embedded Controller Firmware`
* `ChengMing 3967 System Update`
* `Thunderbolt Firmware Update for 0x07E6, XPS9370`
* `Z440/Z640/Z840 Workstation System BIOS`

The translator teams are also frustrated that they cannot translate terms like
`System Update` and also provide the correct versions of `Device`.

To fix, add an enumerated set of firmware 'categories' that can be set by the
uploader in the metainfo.xml file (or changed the LVFS) which automatically
sets the name suffix.

On firmware upload, remove any of the words we're going to suffix and provide
a nag warning to the uploader only. For existing firmware convert them to having
the correct style and fix the names used in the AppStream metadata.

In the future we can set the <category> entry in the metadata and provide
translations in the frontend. This obviously needs changes to fwupd and more
importantly gnome-software and the KDE software center so gently phase this in.