Add an example UEFI firmware for the self tests
This is more complicated than it needs to be to test the PFS-in-Zlib decoding
feature at the same time. The Russian Doll of firmware images now stands at:

* Hello world EFI binary (with banned string in UCS-2)
* AuthentiCode signed binary
* FVH with one single EFI_FFS_FILE_HEADER
* Uncompressed PFS (Dell-specific) image
* ZLib compression
* UEFI Capsule
* Cabinet archive

Some difficult-to-obtain tools like osslsigncode have been used, and so the
resulting archive is also checked into git for the pleasure of CI systems.