Allow firmware or vendors to enable DoNotTrack functionality
This allows either entire vendors to globally opt out of the download and
success reporting, or restrict it from specific firmware objects. Enabling the
LVFS::DoNotTrack feature disables a lot of the useful vendor functionality
(including the per-firmware limits and auto-demotion functionality) and is not
recommended by the LVFS admins.