Allow vendors to disable the .inf parsing
In some cases the .inf file is unparsable, or contains details that are
deliberately different from the values shown when using Windows Update.

To either disable the .inf parsing and verification please use:

    <custom>
      <value key=LVFS::EnableInfParsing>false</value>
    </custom>

The default will remain as checking enabled by default, but firmware should
use `<value key=LVFS::EnableInfParsing>true</value>` if they are relying on the
inf-provided DisplayVersion in case this default changes in the future.