Allow some ODMs to manage the entire firmware cycle
In some cases the OEM is really just a "brand name" and the ODM is responsible
for all phases of uploading, debugging and promoting the firmware update.

For this scenario, allow a different vendor_id_oem key which is used when
we want to show the user information about the OEM rather than the ODM.
In the case where the firmware is actually reassigned to the OEM manually,
the OEM key is updated as well.

This fixes showing various "wrong" ODM vendor logos next to OEM firmware.