Split up the vendor display name
Some OEMs have more than one vendor group on the LVFS. In this case the team
should be displayed in a different way to the OEM name, for instance we never
want to show 'Lenovo ThinkPad ThinkPad P50' when displaying firmware.

We can't migrate this automatically, as quite a few vendors are testing the
LVFS in secret and wouldn't want the internal team names exposed in the
migration scripts.