Include a vendor ID by default for testing accounts
At the moment I create vendor accounts for OEMs and ODMs for testing. I then
create them a first user account, and over time they add more accounts, upload
firmware, test with embargo on real hardware and then ask to be able to promote
to stable.

This default state means that the new user with the @trusted ACL can push
firmware for any vendors hardware as the vendor-id was never set. It's very easy
to forget to ask the vendor what ID to use as part of the onboarding process.

Move to a model where vendors are restricted by default, so that any missing
vendor-ids are caught during testing with embargo.
For OEMs like Dell and Lenovo make the 'unrestricted' vendor ID an explicit
step (with a vendor ID of '*') rather than just a side effect of never
assigning a vendor ID.

When deploying this change we need to remember to manually give the vendor-id
of `*` to all the vendors shipping UEFI capsule or coreboot firmware so that the
vendor ID is not set, as fwupd does not set a vendor-id in all versions.