Reduce the memory consumption when running yara queries
Loading all the thousands of Component objects ahead-of-time means that
sqlalchemy.orm.state.InstanceState starts showing up on profiles using
gigabytes of memory. This makes the small production machine unhappy.

To make it more GC-friendly, load the list of components as integers and then
fetch then ORM objects as required during each query.

This reduces the amount of memory to do a simple yara query on all firmware
from +3Gb to ~300kb.