Add Claim as a top level object
At the moment we just add a key=value attribute to each object that requires a
claim. This doesn't scale, and means we can never change claim icons from
something like 'info' to 'warning' without a database query that touches lots
of records in different tables. We also want to have a URL we can point the
user to for more information, for instance describing what BIOSGuard is,
and why it's a good thing.

To do this, create enumerated claim kinds that can be referenced by ID and
modified as required.