Be explicit when defining python versions
This is good and bad. This means we're less at the mercy of the deployment gods,
as deps won't be able to break API and behaviour at exactly the wrong time.

It also means we version lock, and have to be mindful of fixed security updates
in deps. It also causes pain when we resync with latest and everything breaks
all at once.

On balance, I think I'd rather have a deterministic deployment, and worry less
about if a site upgrade is going to work with slightly different dep versions.