Rewrite the AppStream screenshot URL to use the server CDN
When the screenshot image is loaded on the client system we download the file
from the vendor specified URL. This would cause the client to ping an external
vendor on firmware install, leaking the user agent and IP address.

This probably is not what the user is expecting and certainly not mentioned in
our privacy policy.

Download the image locally when building the metadata, and rewrite the URL in
the public metadata to use the server CDN instead, fixing the privacy issue.