Do not store the client address in the database
We already use the recent nginx for abuse detection, and we don't seem to use
the hashed client address anywhere useful in the LVFS.

The SHA1 hash is not terribly secure (even salted) and removing PII we do not
actually use seems a sane thing to do.

Spotted by Justin Steven <justin@justinsteven.com>, many thanks.